+++ src/OpenWheaterMapChannel.h
@@ -7,7 +7,7 @@ class OpenWheaterMapChannel : public BaseWheaterChannel
   private:
     void fillForecast(JsonObject& json, ForecastWheatherData& wheater);
   protected:
-    int fillWheater(CurrentWheatherData& currentWheater, ForecastWheatherData& todayWheater, ForecastWheatherData& tomorrowWheater) override;
+    int16_t fillWheater(CurrentWheatherData& currentWheater, ForecastWheatherData& todayWheater, ForecastWheatherData& tomorrowWheater) override;
   
   public:
     OpenWheaterMapChannel(uint8_t index);
